requirements.txt
# Jassas Search Engine - Dependencies

# HTML Parsing
beautifulsoup4>=4.12.3
lxml>=5.0.0

# Embeddings (for vector search) - E5 multilingual via FastEmbed
fastembed>=0.4.0

# Vector Index (lightweight, fast)
usearch>=2.9.0

# BM25 Tokenization
nltk>=3.8.0

# API Server
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0

# CLI
typer>=0.13.0

# Utilities
python-dotenv>=1.0.0

requests>=2.32.2            # For crawling web pages
cloudscraper>=1.2.71        # CloudFlare bypass
playwright>=1.40.0          # JS rendering for dynamic sites
numpy==1.26.4               # For math operations
scipy>=1.10.0               # For sparse matrices (BM25 engine)
rich==13.7.1                # For pretty CLI output (The "Manager")